Reorganize Console for bincompat
