--- std/jvm/src/main/scala/cats/effect/std/ConsoleCompanionPlatform.scala
@@ -1,38 +1,0 @@
-/*
- * Copyright 2020-2022 Typelevel
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package cats.effect.std
-
-import cats.effect.kernel.Sync
-import cats.~>
-
-import java.nio.charset.Charset
-
-private[std] trait ConsoleCompanionPlatform { this: Console.type =>
-
-  /**
-   * Constructs a `Console` instance for `F` data types that are [[cats.effect.kernel.Sync]].
-   */
-  def make[F[_]](implicit F: Sync[F]): Console[F] =
-    new SyncConsole[F]
-
-  private[std] abstract class MapKConsole[F[_], G[_]](self: Console[F], f: F ~> G)
-      extends Console[G] {
-    def readLineWithCharset(charset: Charset): G[String] =
-      f(self.readLineWithCharset(charset))
-  }
-
-}
